requirements.txt
keyboard==0.13.5
websockets==12.0
python-rtmidi==1.5.0
